# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EBOS API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/v1/refreshtoken": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Returns JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}